Organise models into groups of Communication, Core and DTO [Processing]
