+++ src/CarCare.Processing/Interfaces/Context/ISessionControlContext.cs
@@ -1,4 +1,4 @@
-﻿using CarCare.Processing.Models;
+﻿using CarCare.Processing.Models.Dto;
 using System.Windows;
 using System.Windows.Input;
