+++ src/CarCare.Processing/Interfaces/Session/IUserSession.cs
@@ -1,4 +1,5 @@
-﻿using CarCare.Processing.Models;
+﻿using CarCare.Processing.Models.Core;
+using CarCare.Processing.Models.Dto;
 
 namespace CarCare.Processing.Interfaces.Session;
